% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/21_mcmc_alpha.R
\name{hs_alpha_gibbs}
\alias{hs_alpha_gibbs}
\title{Horseshoe-Gibbs for synthetic weights alpha}
\usage{
hs_alpha_gibbs(y, X, M = 2000, burn = 1000, verbose = TRUE)
}
\description{
Horseshoe-Gibbs for synthetic weights alpha
}
\keyword{internal}
