% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.R
\name{posterior_effects}
\alias{posterior_effects}
\title{Posterior effects via identification formulas (5)(6)}
\usage{
posterior_effects(Y0_post, Yc_post, alpha_draws, rho_draws, w, W, cred = 0.95)
}
\description{
Posterior effects via identification formulas (5)(6)
}
\keyword{internal}
