% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/22_mcmc_sar.R
\name{sar_gibbs_sampler}
\alias{sar_gibbs_sampler}
\title{Full SAR MCMC (rho, beta, sigma2, Lambda, F) with optional covariates and factors}
\usage{
sar_gibbs_sampler(
  Y0_pre,
  Yc_pre,
  Xc_pre,
  W,
  w,
  M = 2000,
  burn = 1000,
  verbose = TRUE,
  p_factors = 0,
  step_rho = 0.05,
  c_beta = 10,
  c_lambda = 10,
  a0 = 1,
  b0 = 1
)
}
\description{
Full SAR MCMC (rho, beta, sigma2, Lambda, F) with optional covariates and factors
}
\keyword{internal}
