% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_utils_data_prep.R
\name{scspill_prep}
\alias{scspill_prep}
\title{Prepare Y matrices from long panel}
\usage{
scspill_prep(data, treated_unit, T0)
}
\description{
Prepare Y matrices from long panel
}
\keyword{internal}
