% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.R
\name{scspill_prep_X}
\alias{scspill_prep_X}
\title{Prepare panel & covariates into matrices (controls only)}
\usage{
scspill_prep_X(
  data,
  treated_unit,
  T0,
  X = NULL,
  y_col = "y",
  unit_col = "unit",
  time_col = "time"
)
}
\description{
Prepare panel & covariates into matrices (controls only)
}
\keyword{internal}
